pyproject.toml
[project]
name = "WiseCal"
version = "0.1.0"
description = "Sync Wise Timetable to Google Calendar"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.1",
    "flask>=3.1.2",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "icalendar>=6.3.2",
    "playwright>=1.55.0",
    "pyyaml>=6.0.3",
    "waitress>=3.0.2",
]
